Start to work on library, changes to build.py
